# benchmark: use temporary file for Get/Put object benchmark tests.