Fixes #2132