@hackintoshrao 2016-07-10
@balamurugana : Can you run benchcmp for the benchmarks run for the code with your changes in setup and with the current master setup and see the allocations difference in percentage. That'll help us understand how significant these changes are.