@balamurugana 2016-07-11
Below is result of benchcmp output.

```bash
$ benchcmp bench1 bench2
benchmark                      old ns/op     new ns/op     delta
BenchmarkPutObject25MbFS-8     255471092     259878610     +1.73%
BenchmarkPutObject25MbXL-8     430762868     421368596     -2.18%

benchmark                      old allocs     new allocs     delta
BenchmarkPutObject25MbFS-8     3451           3446           -0.14%
BenchmarkPutObject25MbXL-8     2883           2875           -0.28%

benchmark                      old bytes     new bytes     delta
BenchmarkPutObject25MbFS-8     379955        379444        -0.13%
BenchmarkPutObject25MbXL-8     116081181     116074562     -0.01%
```

I also did memprofile individually for FS and XL test in current master.

```bash
$ go test -v -run=NONE -bench BenchmarkPutObject25MbFS -memprofile=fs.mem && go tool pprof minio.test fs.mem
PASS
BenchmarkPutObject25MbFS-8             5         255158351 ns/op          379939 B/op       3451 allocs/op
ok      github.com/minio/minio  2.962s
Entering interactive mode (type "help" for commands)
(pprof) top --cum
512.02kB of 512.02kB total (  100%)
Dropped 14 nodes (cum <= 2.56kB)
      flat  flat%   sum%        cum   cum%
         0     0%     0%   512.02kB   100%  github.com/minio/minio.init
  512.02kB   100%   100%   512.02kB   100%  internal/golang.org/x/net/http2/hpack.addDecoderNode
         0     0%   100%   512.02kB   100%  internal/golang.org/x/net/http2/hpack.init
         0     0%   100%   512.02kB   100%  internal/golang.org/x/net/http2/hpack.init.1
         0     0%   100%   512.02kB   100%  main.init
         0     0%   100%   512.02kB   100%  net/http.init
         0     0%   100%   512.02kB   100%  runtime.goexit
         0     0%   100%   512.02kB   100%  runtime.main
(pprof) top10
512.02kB of 512.02kB total (  100%)
Dropped 14 nodes (cum <= 2.56kB)
      flat  flat%   sum%        cum   cum%
  512.02kB   100%   100%   512.02kB   100%  internal/golang.org/x/net/http2/hpack.addDecoderNode
         0     0%   100%   512.02kB   100%  github.com/minio/minio.init
         0     0%   100%   512.02kB   100%  internal/golang.org/x/net/http2/hpack.init
         0     0%   100%   512.02kB   100%  internal/golang.org/x/net/http2/hpack.init.1
         0     0%   100%   512.02kB   100%  main.init
         0     0%   100%   512.02kB   100%  net/http.init
         0     0%   100%   512.02kB   100%  runtime.goexit
         0     0%   100%   512.02kB   100%  runtime.main
(pprof) 

$ go test -v -run=NONE -bench BenchmarkPutObject25MbXL -memprofile=xl.mem && go tool pprof minio.test xl.mem
PASS
BenchmarkPutObject25MbXL-8             3         424817304 ns/op        116078888 B/op      2881 allocs/op
ok      github.com/minio/minio  3.465s
Entering interactive mode (type "help" for commands)
(pprof) top --cum
150MB of 150MB total (  100%)
Dropped 57 nodes (cum <= 0.75MB)
      flat  flat%   sum%        cum   cum%
         0     0%     0%      150MB   100%  github.com/minio/minio.(*xlObjects).PutObject
         0     0%     0%      150MB   100%  github.com/minio/minio.BenchmarkPutObject25MbXL
         0     0%     0%      150MB   100%  github.com/minio/minio.benchmarkPutObject
         0     0%     0%      150MB   100%  github.com/minio/minio.returnPutObjectBenchmark.func1
         0     0%     0%      150MB   100%  github.com/minio/minio.runPutObjectBenchmark
         0     0%     0%      150MB   100%  github.com/minio/minio.xlObjects.PutObject
     150MB   100%   100%      150MB   100%  github.com/minio/minio/pkg/objcache.(*Cache).Create
         0     0%   100%      150MB   100%  runtime.goexit
         0     0%   100%      150MB   100%  testing.(*B).launch
         0     0%   100%      150MB   100%  testing.(*B).runN
(pprof) top10
150MB of 150MB total (  100%)
Dropped 57 nodes (cum <= 0.75MB)
      flat  flat%   sum%        cum   cum%
     150MB   100%   100%      150MB   100%  github.com/minio/minio/pkg/objcache.(*Cache).Create
         0     0%   100%      150MB   100%  github.com/minio/minio.(*xlObjects).PutObject
         0     0%   100%      150MB   100%  github.com/minio/minio.BenchmarkPutObject25MbXL
         0     0%   100%      150MB   100%  github.com/minio/minio.benchmarkPutObject
         0     0%   100%      150MB   100%  github.com/minio/minio.returnPutObjectBenchmark.func1
         0     0%   100%      150MB   100%  github.com/minio/minio.runPutObjectBenchmark
         0     0%   100%      150MB   100%  github.com/minio/minio.xlObjects.PutObject
         0     0%   100%      150MB   100%  runtime.goexit
         0     0%   100%      150MB   100%  testing.(*B).launch
         0     0%   100%      150MB   100%  testing.(*B).runN
(pprof) 
```

As object caching is involved, previous `bytes.Repeat()` memory consumption is not seen.

Aborting this patch.